It wasn't a bug. It just wasn't implemented. (WIP)

+++ core/src/main/dev/kabin/utilities/Functions.java
@@ -239,7 +239,7 @@ public class Functions {
         return (x, y) -> minX <= x && x < maxX && minY <= y && y < maxY;
     }
 
-    public static int toInt(double d, double scale) {
+    public static int toInt(float d, float scale) {
         return (int) Math.round(d / scale);
     }
